fixes for memory/stack values reading
* fix reg<->bignum conversion (always little-endian)
* fix RETURN op (memory reading)
* fix PUSH op stack reading
* fix MLOAD, MSTORE (write straight bytes, read reversed)
* fix reading addr_val from stack (memory ops)
* fix MSTORE8 (getting byte)
+ add BYTE op realization

+++ emulator/models/evm/simulator/evm_sim.h
@@ -205,7 +205,7 @@ inline uint64 DelegateCall();
 inline uint64 PushInst(uint32 cnt);
 inline uint64 LogInst(uint32 count,uint32 log_ptr);
 inline uint64 MloadInst(uint32 addr_val);
-inline uint64 MStoreInst(uint32 addr_val);
+inline uint64 MStoreInst(uint64 addr_val);
   inline int Main_decode(uint32 ocode);
   inline int EVM_Main_GetInstNum(uint32 ocode);
   void update();
